refactor: put footer inside general components
