+++ components/gernal/footer.tsx
